+++ src/Subscriptions.tsx
@@ -1,11 +1,12 @@
-import * as React from 'react';
-import * as PropTypes from 'prop-types';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ApolloClient, { ApolloError } from 'apollo-client';
 import { Observable } from 'apollo-link';
 
 import { DocumentNode } from 'graphql';
 import { ZenObservable } from 'zen-observable-ts';
 import { OperationVariables } from './types';
+import { getClient } from './component-utils';
 
 const shallowEqual = require('fbjs/lib/shallowEqual');
 const invariant = require('invariant');
@@ -16,11 +17,18 @@ export interface SubscriptionResult<TData = any> {
   error?: ApolloError;
 }
 
+export interface OnSubscriptionDataOptions<TData = any> {
+  client: ApolloClient<Object>;
+  subscriptionData: SubscriptionResult<TData>;
+}
+
 export interface SubscriptionProps<TData = any, TVariables = OperationVariables> {
   subscription: DocumentNode;
   variables?: TVariables;
   shouldResubscribe?: any;
-  children: (result: SubscriptionResult<TData>) => React.ReactNode;
+  client?: ApolloClient<Object>;
+  onSubscriptionData?: (options: OnSubscriptionDataOptions<TData>) => any;
+  children?: (result: SubscriptionResult<TData>) => React.ReactNode;
 }
 
 export interface SubscriptionState<TData = any> {
@@ -30,7 +38,7 @@ export interface SubscriptionState<TData = any> {
 }
 
 export interface SubscriptionContext {
-  client: ApolloClient<Object>;
+  client?: ApolloClient<Object>;
 }
 
 class Subscription<TData = any, TVariables = any> extends React.Component<
@@ -44,7 +52,8 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   static propTypes = {
     subscription: PropTypes.object.isRequired,
     variables: PropTypes.object,
-    children: PropTypes.func.isRequired,
+    children: PropTypes.func,
+    onSubscriptionData: PropTypes.func,
     shouldResubscribe: PropTypes.oneOfType([PropTypes.func, PropTypes.bool]),
   };
 
@@ -55,11 +64,7 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   constructor(props: SubscriptionProps<TData, TVariables>, context: SubscriptionContext) {
     super(props, context);
 
-    invariant(
-      !!context.client,
-      `Could not find "client" in the context of Subscription. Wrap the root component in an <ApolloProvider>`,
-    );
-    this.client = context.client;
+    this.client = getClient(props, context);
     this.initialize(props);
     this.state = this.getInitialState();
   }
@@ -72,9 +77,11 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
     nextProps: SubscriptionProps<TData, TVariables>,
     nextContext: SubscriptionContext,
   ) {
+    const nextClient = getClient(nextProps, nextContext);
+
     if (
       shallowEqual(this.props.variables, nextProps.variables) &&
-      this.client === nextContext.client &&
+      this.client === nextClient &&
       this.props.subscription === nextProps.subscription
     ) {
       return;
@@ -85,8 +92,8 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
       shouldResubscribe = !!shouldResubscribe(this.props, nextProps);
     }
     const shouldNotResubscribe = shouldResubscribe === false;
-    if (this.client !== nextContext.client) {
-      this.client = nextContext.client;
+    if (this.client !== nextClient) {
+      this.client = nextClient;
     }
 
     if (!shouldNotResubscribe) {
@@ -106,10 +113,12 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   }
 
   render() {
+    const renderFn: any = this.props.children;
+    if (!renderFn) return null;
     const result = Object.assign({}, this.state, {
       variables: this.props.variables,
     });
-    return this.props.children(result);
+    return renderFn(result);
   }
 
   private initialize = (props: SubscriptionProps<TData, TVariables>) => {
@@ -135,6 +144,11 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   });
 
   private updateCurrentData = (result: SubscriptionResult<TData>) => {
+    const {
+      client,
+      props: { onSubscriptionData },
+    } = this;
+    if (onSubscriptionData) onSubscriptionData({ client, subscriptionData: result });
     this.setState({
       data: result.data,
       loading: false,